chore(rules): moves rules related things into the rules module
