+++ src/app/connections/views/ConnectionOutboundSummaryOverviewView.vue
@@ -35,8 +35,8 @@
           >
             <h3>Rules</h3>
             <DataLoader
-              v-slot="{ data: rulesData }: DataplaneRulesSource"
-              :src="`/meshes/${route.params.mesh}/dataplanes/${route.params.dataPlane}/rules`"
+              v-slot="{ data: rulesData }: RuleCollectionSource"
+              :src="`/meshes/${route.params.mesh}/rules/for/${route.params.dataPlane}`"
             >
               <DataCollection
                 v-slot="{ items }"
@@ -100,7 +100,7 @@
                                   <template #body>
                                     <DataSource
                                       v-slot="{ data: policyTypes }: PolicyTypeCollectionSource"
-                                      :src="`/*/policy-types`"
+                                      :src="`/policy-types`"
                                     >
                                       <template
                                         v-for="types in [Object.groupBy((policyTypes?.policies ?? []), (item) => item.name)]"
@@ -172,9 +172,9 @@ import CodeBlock from '@/app/common/code-block/CodeBlock.vue'
 import DefinitionCard from '@/app/common/DefinitionCard.vue'
 import PolicyTypeTag from '@/app/common/PolicyTypeTag.vue'
 import type { DataplaneOverview } from '@/app/data-planes/data/'
-import type { DataplaneRulesSource } from '@/app/data-planes/sources'
 import type { PolicyTypeCollectionSource } from '@/app/policies/sources'
 import RuleMatchers from '@/app/rules/components/RuleMatchers.vue'
+import type { RuleCollectionSource } from '@/app/rules/sources'
 import { toYaml } from '@/utilities/toYaml'
 const props = defineProps<{
   data: Record<string, any>
