+++ src/app/data-planes/locales/en-us/index.yaml
@@ -63,9 +63,10 @@ data-planes:
           This Data Plane Proxy does not have mTLS configured, yet  — <a href="{KUMA_DOCS_URL}/policies/mutual-tls?{KUMA_UTM_QUERY_PARAMS}">Learn about certificates in {KUMA_PRODUCT_NAME}</a>
       subscriptions:
         title: XDS Configurations
-      proxy_rule: Proxy rule
-      to_rules: To rules
-      from_rules: From rules
+      rules:
+        proxy: Proxy rule
+        to: To rules
+        from: From rules
       legacy_policies: Legacy policies
       matches_everything: 'Matches everything'
     items:
