+++ src/app/gateways/sources.ts
@@ -1,7 +1,7 @@
 import { MeshGateway } from './data'
 import type { DataSourceResponse } from '@/app/application'
 import { defineSources } from '@/app/application/services/data-source'
-import { Rule, type RuleCollection } from '@/app/data-planes/data'
+import { Rule, type RuleCollection } from '@/app/rules/data'
 import type KumaApi from '@/services/kuma-api/KumaApi'
 import type { PaginatedApiListResponse as CollectionResponse } from '@/types/api.d'
