+++ src/app/rules/index.ts
@@ -1,0 +1,20 @@
+import { sources } from './sources'
+import type { ServiceDefinition } from '@/services/utils'
+import { token } from '@/services/utils'
+
+type Token = ReturnType<typeof token>
+
+export const services = (app: Record<string, Token>): ServiceDefinition[] => {
+  return [
+    [token('rules.sources'), {
+      service: sources,
+      arguments: [
+        app.source,
+        app.api,
+      ],
+      labels: [
+        app.sources,
+      ],
+    }],
+  ]
+}
